+++ workspaces/bulk-import/plugins/bulk-import/src/components/GitLabIcon.tsx
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import * as React from 'react';
+import type { FC, HTMLProps, ReactElement } from 'react';
 
-const GitLabIcon: React.FC<React.HTMLProps<SVGElement>> = ({
-  style,
-}): React.ReactElement => {
+const GitLabIcon: FC<HTMLProps<SVGElement>> = ({ style }): ReactElement => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"