+++ workspaces/sandbox/plugins/sandbox/src/components/Modals/PhoneVerificationSteps/VerificationCodeStep.tsx
@@ -62,6 +62,7 @@ export const VerificationCodeStep: React.FC<VerificationCodeProps> = ({
   const inputRefs = useRef<any>([]);
 
   useEffect(() => {
+    // Focus on the first input box when modal opens
     if (!otp[0]) {
       // Focus on the first input box when modal opens
       setTimeout(() => inputRefs.current[0]?.focus(), 100);
@@ -115,6 +116,7 @@ export const VerificationCodeStep: React.FC<VerificationCodeProps> = ({
         Enter the verification code
       </DialogTitle>
       <IconButton
+        data-testid="close-opt-button"
         aria-label="close"
         onClick={handleClose}
         sx={{
@@ -128,6 +130,7 @@ export const VerificationCodeStep: React.FC<VerificationCodeProps> = ({
       </IconButton>
       <DialogContent sx={{ padding: '6px 24px' }}>
         <DialogContentText
+          data-testid="sent-message-dialog"
           id="alert-dialog-description"
           color="textPrimary"
           style={{
@@ -151,6 +154,7 @@ export const VerificationCodeStep: React.FC<VerificationCodeProps> = ({
             {otp.map((digit, index) => (
               <TextField
                 key={index}
+                data-testid="opt-inputs"
                 value={digit}
                 onChange={e => handleChange(index, e)}
                 onKeyDown={e =>
@@ -171,6 +175,7 @@ export const VerificationCodeStep: React.FC<VerificationCodeProps> = ({
         </div>
 
         <Typography
+          data-testid="resend-code-link"
           variant="body2"
           color="primary"
           sx={{
@@ -211,10 +216,11 @@ export const VerificationCodeStep: React.FC<VerificationCodeProps> = ({
       </DialogContent>
       <DialogActions sx={{ justifyContent: 'flex-start', padding: '24px' }}>
         <Button
+          data-testid="submit-opt-button"
           variant="contained"
           type="submit"
           onClick={handleStartTrialClick}
-          disabled={otp.some(digit => !digit) || loading || !!error}
+          disabled={otp.some(digit => !digit) || loading}
           endIcon={
             loading && <CircularProgress size={20} sx={{ color: '#AFAFAF' }} />
           }
@@ -222,6 +228,7 @@ export const VerificationCodeStep: React.FC<VerificationCodeProps> = ({
           Start trial
         </Button>
         <Button
+          data-testid="close-opt-button"
           variant="outlined"
           onClick={handleClose}
           sx={{