+++ workspaces/adoption-insights/plugins/adoption-insights-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @red-hat-developer-hub/backstage-plugin-adoption-insights-common
 
+## 0.2.0
+
+### Minor Changes
+
+- a66e0b0: Backstage version bump to v1.36.1
+
 ## 0.1.1
 
 ### Patch Changes