+++ workspaces/orchestrator/plugins/orchestrator-backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 ### Dependencies
 
+## 6.0.2
+
+### Patch Changes
+
+- 7f6ca8a: add filter by variables and nested variables
+- Updated dependencies [c6b54ad]
+- Updated dependencies [7f6ca8a]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@1.27.2
+
+## 6.0.1
+
+### Patch Changes
+
+- 56c160b: Add endpoind to ping workflow service
+- 571d93e: Updated dependency `@types/express` to `4.17.22`.
+- Updated dependencies [56c160b]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@1.27.1
+
 ## 6.0.0
 
 ### Major Changes