+++ workspaces/orchestrator/plugins/scaffolder-backend-module-orchestrator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @red-hat-developer-hub/backstage-plugin-scaffolder-backend-module-orchestrator
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [c6b54ad]
+- Updated dependencies [7f6ca8a]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@1.27.2
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [56c160b]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@1.27.1
+
 ## 0.3.0
 
 ### Minor Changes